No Red, no semantics, only syntax.
Honestly, I couldn't figure out how to make this simple parser in Red in the few days that I had (returning the AST is the hard part), and I wanted to implement (think through) semantics in semantics first (and transliterate them later, definitely in JS, possibly in Red), so… No Red, no semantics, only syntax.

There is still `war.md`, but it has even less interesting things in it. Semantics are too incomplete.