Made {…}/[…] work properly.
Which is no good without functions or any/all. Or, you know, all of those.

But it helps.

(The rest of semantic rules specified inside is a jumbled mess.)

---

If `0|1` is the type of bits, then `0&1` is the type of a superposition of bits. When bound to, `0|1` succeeds if any part succeeds (much like `{0 1}`, but done in a random order, returning any success); `0&1` succeeds if all parts succeed (which can only happen for superpositions) (much like `[0 1]`, but done in a random order, returning all successes).

0|1 = 0
0|1 = 1
0&1 = 0&1

`A|B` can be used when it is known/assumed that A and B are equivalent; `A&B`... is separate from that notion.

Can `all` replace `finish`; is one the other's potentially-parallel version?